requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
google-generativeai
